ajout des différents types de briques

+++ src/Platform.cpp
@@ -7,7 +7,9 @@ void Platform::setSpeed(int speed) { _speed = speed; }
 void Platform::draw(SDL_Renderer *renderer) {
     SDL_SetRenderDrawColor(renderer, getColor().r, getColor().g, getColor().b,
                            getColor().a);
-    SDL_Rect rect = {getX(), getY(), getWidth(), getHeight()};
+
+    SDL_Rect rect = {(int)(getX()), (int)getY(), (int)getWidth(),
+                     (int)getHeight()};
     SDL_RenderFillRect(renderer, &rect);
 }
